BTS count based on Poisson distribution

+++ src/com/lteu/deliverable/Location.java
@@ -12,7 +12,7 @@ public class Location {
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
@@ -20,7 +20,7 @@ public class Location {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
